+++ Toolchain/Root.zig
@@ -21,4 +21,6 @@ test {
     _ = @import("Sources/MacOS/Link.zig");
     _ = @import("Sources/Llvm/Units.zig");
     _ = @import("Sources/Llvm/Store.zig");
+    _ = @import("Sources/Arm64/SystemHeap.zig");
+    _ = @import("Sources/X64/SystemHeap.zig");
 }